pyproject.toml
[project]
name = "multimodal-CLIP-demo"
version = "2026.2.8"
readme = "README.md"
license = "Apache-2.0"

dependencies = [
  "typing",
  "fastapi[standard]",
  "jinja2",
  "pillow",
  "psycopg[binary]",
  "python-dotenv",
  "torch",
  "huggingface_hub",
  #"transformers",
  # Version 5 of the HuggingFace transformers package introduced a
  # breaking change to the output of get_image_features. For now
  # stick with an older version.
  "transformers==4.57.3",  # try an old version of transformers

]

[tool.setuptools.packages]
# We've just got our app.py at the top level, nothing to look for
find = {namespaces = false}
